refactor(voyager): camelCase remaining kebab dir and file names (R-237, R-217)
Renames enrichment-sources/ to enrichmentSources/, tool-call-log.ts to toolCallLog.ts, and visa-matrix.ts to visaMatrix.ts (plus test mirrors); updates all imports and the coverage exclude list.
